[eas-cli] swap beta credentials manager with current
